Make SparklesImage a shared component
